Zakaria/keyboost_web_app/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cpu
statsmodels
sklearn
scipy
streamlit
pandas
numpy
Pillow
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0.tar.gz#egg=en_core_web_sm
https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.0.0/fr_core_news_sm-3.0.0.tar.gz#egg=fr_core_news_sm
yake
keybert
gensim==3.8.3
